Added TTL trigger out synced to stimulus
Added .rcx file changes and a few lines in Matlab code to produce a trigger leaving the RZ6 DB25 every time sound or laser is activated. The trigger is configured to match `acq_samples` under the hood, which is derived from the user-defined variable `acquire_length_ms` in the GUI.

The trigger is on for the _recording_ duration of a stimulus. In our experiments, this is typically longer than the acoustic/laser stimulus duration itself. For example, if recording 100ms of sound + 150ms of silence, `acquire_length_ms` will be 250 ms. The goal is for the ITC-18 & Igor to match the legacy matlab recording setup in its recording duration. This trigger supplies that information from the matlab/RZ6 end.